// src/types/OmnivoreSchema.ts
export type Maybe<T> = T | null
export type Exact<T extends Record<string, unknown>> = {
  [K in keyof T]: T[K]
}
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>
}
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>
}
/** All built-in and custom scalars, mapped to their actual values */
export interface Scalars {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  Date: Date
}
// ...
export interface Label {
  __typename?: 'Label'
  id: Scalars['ID']
  name: Scalars['String']
  color?: Scalars['String']
  description?: Maybe<Scalars['String']>
  createdAt?: Maybe<Scalars['Date']>
  position?: Maybe<Scalars['Int']>
  internal?: Maybe<Scalars['Boolean']>
}